refactor: migrate data store interfaces to the org.groundplatform.data.stores package
