+++ app/src/main/java/org/groundplatform/android/data/local/room/stores/RoomUserStore.kt
@@ -22,7 +22,7 @@ import org.groundplatform.android.data.local.room.converter.toLocalDataStoreObje
 import org.groundplatform.android.data.local.room.converter.toModelObject
 import org.groundplatform.android.data.local.room.dao.UserDao
 import org.groundplatform.android.data.local.room.dao.insertOrUpdate
-import org.groundplatform.android.data.local.stores.LocalUserStore
+import org.groundplatform.data.stores.LocalUserStore
 import org.groundplatform.domain.model.User
 
 /** Manages access to [User] objects persisted in local storage. */
