+++ app/src/main/java/org/groundplatform/android/repository/SurveyRepository.kt
@@ -33,9 +33,9 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.withTimeout
 import org.groundplatform.android.FirebaseCrashLogger
 import org.groundplatform.android.data.local.LocalValueStore
-import org.groundplatform.android.data.local.stores.LocalSurveyStore
-import org.groundplatform.android.data.remote.RemoteDataStore
 import org.groundplatform.android.di.coroutines.ApplicationScope
+import org.groundplatform.data.stores.LocalSurveyStore
+import org.groundplatform.data.stores.RemoteDataStore
 import org.groundplatform.domain.model.Survey
 import org.groundplatform.domain.model.SurveyListItem
 import org.groundplatform.domain.model.User
