+++ app/src/main/java/org/groundplatform/android/repository/TermsOfServiceRepository.kt
@@ -21,7 +21,7 @@ import javax.inject.Singleton
 import kotlinx.coroutines.withTimeout
 import org.groundplatform.android.data.local.LocalValueStore
 import org.groundplatform.android.data.remote.DataStoreException
-import org.groundplatform.android.data.remote.RemoteDataStore
+import org.groundplatform.data.stores.RemoteDataStore
 import org.groundplatform.domain.model.TermsOfService
 import org.groundplatform.domain.repository.TermsOfServiceRepositoryInterface
 import org.groundplatform.domain.system.NetworkManagerInterface
