+++ app/src/main/java/org/groundplatform/android/repository/UserRepository.kt
@@ -20,9 +20,9 @@ import javax.inject.Singleton
 import kotlinx.coroutines.flow.Flow
 import org.groundplatform.android.data.local.LocalValueStore
 import org.groundplatform.android.data.local.room.LocalDatabase
-import org.groundplatform.android.data.local.stores.LocalUserStore
-import org.groundplatform.android.data.remote.RemoteDataStore
 import org.groundplatform.android.system.auth.AuthenticationManager
+import org.groundplatform.data.stores.LocalUserStore
+import org.groundplatform.data.stores.RemoteDataStore
 import org.groundplatform.domain.model.Role
 import org.groundplatform.domain.model.Survey
 import org.groundplatform.domain.model.User
