+++ app/src/test/java/org/groundplatform/android/data/local/LocalOfflineAreaStore.kt
@@ -20,7 +20,7 @@ import com.google.common.truth.Truth.assertThat
 import dagger.hilt.android.testing.HiltAndroidTest
 import javax.inject.Inject
 import org.groundplatform.android.BaseHiltTest
-import org.groundplatform.android.data.local.stores.LocalOfflineAreaStore
+import org.groundplatform.data.stores.LocalOfflineAreaStore
 import org.groundplatform.domain.model.imagery.OfflineArea
 import org.groundplatform.domain.model.map.Bounds
 import org.junit.Test
