+++ app/src/test/java/org/groundplatform/android/data/local/LocalUserStoreTest.kt
@@ -20,7 +20,7 @@ import dagger.hilt.android.testing.HiltAndroidTest
 import javax.inject.Inject
 import org.groundplatform.android.BaseHiltTest
 import org.groundplatform.android.FakeData
-import org.groundplatform.android.data.local.stores.LocalUserStore
+import org.groundplatform.data.stores.LocalUserStore
 import org.groundplatform.domain.model.User
 import org.junit.Test
 import org.junit.runner.RunWith
