+++ app/src/test/java/org/groundplatform/android/data/remote/FakeRemoteDataStore.kt
@@ -20,6 +20,7 @@ import javax.inject.Singleton
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOf
+import org.groundplatform.data.stores.RemoteDataStore
 import org.groundplatform.domain.model.Survey
 import org.groundplatform.domain.model.SurveyListItem
 import org.groundplatform.domain.model.TermsOfService
