+++ app/src/test/java/org/groundplatform/android/repository/SubmissionRepositoryTest.kt
@@ -18,9 +18,9 @@ package org.groundplatform.android.repository
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.test.runTest
 import org.groundplatform.android.data.local.LocalValueStore
-import org.groundplatform.android.data.local.stores.LocalSubmissionStore
 import org.groundplatform.android.data.sync.MutationSyncWorkManager
 import org.groundplatform.android.data.uuid.OfflineUuidGenerator
+import org.groundplatform.data.stores.LocalSubmissionStore
 import org.groundplatform.domain.model.User
 import org.groundplatform.domain.model.locationofinterest.AuditInfo
 import org.groundplatform.domain.model.locationofinterest.LocationOfInterest
