+++ app/src/test/java/org/groundplatform/android/ui/offlineareas/OfflineAreasFragmentTest.kt
@@ -38,8 +38,8 @@ import kotlinx.coroutines.test.advanceUntilIdle
 import org.groundplatform.android.BaseHiltTest
 import org.groundplatform.android.FakeData.OFFLINE_AREA
 import org.groundplatform.android.R
-import org.groundplatform.android.data.local.stores.LocalOfflineAreaStore
 import org.groundplatform.android.testrules.FragmentScenarioRule
+import org.groundplatform.data.stores.LocalOfflineAreaStore
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
