+++ app/src/test/java/org/groundplatform/android/ui/offlineareas/viewer/OfflineAreaViewerFragmentTest.kt
@@ -32,9 +32,9 @@ import kotlinx.coroutines.test.advanceUntilIdle
 import org.groundplatform.android.BaseHiltTest
 import org.groundplatform.android.FakeData.OFFLINE_AREA
 import org.groundplatform.android.R
-import org.groundplatform.android.data.local.stores.LocalOfflineAreaStore
 import org.groundplatform.android.testrules.FragmentScenarioRule
 import org.groundplatform.android.util.view.isGone
+import org.groundplatform.data.stores.LocalOfflineAreaStore
 import org.groundplatform.domain.model.map.MapType
 import org.groundplatform.ui.map.MapConfig
 import org.junit.Rule
