+++ app/src/test/java/org/groundplatform/android/ui/offlineareas/viewer/OfflineAreaViewerViewModelTest.kt
@@ -21,7 +21,7 @@ import javax.inject.Inject
 import kotlinx.coroutines.test.advanceUntilIdle
 import org.groundplatform.android.BaseHiltTest
 import org.groundplatform.android.FakeData.OFFLINE_AREA
-import org.groundplatform.android.data.local.stores.LocalOfflineAreaStore
+import org.groundplatform.data.stores.LocalOfflineAreaStore
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
