+++ core/data/src/commonMain/kotlin/org/groundplatform/data/stores/LocalMutationStore.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.groundplatform.android.data.local.stores
+package org.groundplatform.data.stores
 
 import org.groundplatform.domain.model.mutation.Mutation
